requirements.txt
Django==1.9
djangorestframework==3.0.0
drf-nested-routers==0.11.1
